Update protectintegrationcreate.sh
The echo on the access token has likely already been removed in the staging branch. I made my sonar-fix branch before that happened but  hopefully we won't get a merge conflict.

There's no good way to construct a JSON body in Bash. the linter did not like that the variables in the data parameter were not double-quoted because it will cause globbing errors if the value contains a space or quote. They won't, of course, but the linter doesn't know that.

jq is the better way to do JSON since it does escaping really nice, but jq was only recently added to macOS. So my suggestion is not the right way to do it if we need to support older OS versions.

This would be how to do it without jq and will work with older OSs. But it's pretty hard to read...

data='{"protectUrl": "'"$jamfprotect_url"'","clientId": "'"$jamfprotect_clientID"'","password": "'"$jamfprotect_client_password"'"}'